Fix a part of the code that was cheating
Previously, we were memoizing the "longest substring" calculation, but
we were storing this information *between runs* of the `sort` method,
which is cheating, since later runs of the sort method will perform
better.

The fix is to create an array of `Data` every time the `sort` method
gets called, instead of once at the beginning of the program.

Sorry about that!